pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ttkbootstrap"
version = "1.13.5"
description = "A supercharged theme extension for tkinter that enables on-demand modern flat style themes inspired by Bootstrap."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name="Israel Dryer", email="israel.dryer@gmail.com" }
]
keywords = [
    "tkinter",
    "theme",
    "ttk",
    "bootstrap",
    "gui",
    "desktop",
    "python3",
    "ui",
    "widgets",
    "theming",
    "cross-platform",
    "modern",
    "responsive",
    "python-gui"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: User Interfaces",
    "Operating System :: OS Independent",
]

dependencies = [
    "pillow>=10,<11",
]

[project.urls]
Homepage = "https://github.com/israel-dryer/ttkbootstrap"
Documentation = "https://ttkbootstrap.readthedocs.io/en/latest/"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ttkbootstrap" = ["localization/msgs/*.msg", "themes/*.json", "assets/*"]
